Create functionality for the yes button to load into the new expense page
