+++ client/src/App.tsx
@@ -1,8 +1,9 @@
 import { Route, Routes, Navigate } from 'react-router-dom';
 import { AuthPage } from './pages/SwitchPage';
 import { UserProvider } from './components/UserContext';
-import { RegistrationForm } from './components/SignUp';
+import { RegistrationForm } from './pages/SignUp';
 import { Home } from './pages/Home';
+import { NewExpense } from './pages/NewExpense';
 
 export function App() {
   return (
@@ -11,6 +12,7 @@ export function App() {
         <Route index element={<RegistrationForm />} />
         <Route path="/sign-up" element={<Navigate to="/" />} />
         <Route path="/log-in" element={<AuthPage mode="log-in" />} />
+        <Route path="/new-expense" element={<NewExpense />} />
         <Route path="/home" element={<Home />} />
       </Routes>
     </UserProvider>
